requirements.txt
numpy~=1.21.5
scipy~=1.7.3
matplotlib~=3.4.3
scikit-learn~=1.0.1
pandas~=1.3.4
joblib~=1.1.0
seaborn~=0.11.0
pyyaml~=6.0
optuna~=2.10.0
torch~=1.10.0
pytorch-lightning~=1.5.1
POT~=0.8.2
neptune-client~=0.15.2